o integrate VNPAY payment into the website

+++ e-shop/app/Http/Controllers/Frontend/CheckoutController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use  Gloudemans\Shoppingcart\Facades\Cart;
 use App\Models\Order;
 use App\Models\OrderDetail;
+use App\Utilities\VNPay;
 class CheckoutController extends Controller
 {
     //
@@ -30,13 +31,43 @@ class CheckoutController extends Controller
                 'total'=> $item->price * $item->qty
             ]);
         }
-        //Xóa giỏ hàng
-        Cart::destroy();
-        //Trả về kết quả
-        return redirect('checkout/result')->with('success','Cảm ơn bạn đã đặt hàng vui lòng kiểm tra email!');
+        if($request->payment=='pay_later'){
+             //Xóa giỏ hàng
+            Cart::destroy();
+            //Trả về kết quả
+            return redirect('checkout/result')->with('notice','Cảm ơn bạn đã đặt hàng vui lòng kiểm tra email!');
+        }
+        if($request->payment=='online_pay'){
+            //Lấy url thanh toán VN Pay
+            $data_url= VNPay::vnpay_create_payment([
+                'vnp_TxnRef' => $order->id,
+                'vnp_OrderInfo' => 'Đơn hàng',
+                'vnp_OrderType' => 'billpayment',
+                'vnp_Amount' => Cart::total(0,'','.')*23000, // giá trị $ sang VND
+            ]);
+            return redirect($data_url);
+        }
+    }
+    public function vnPayCheck(Request $request){
+        //Lấy data từ url (VN Pay gửi cho qua $vnp_Returnurl)
+        $vnp_ReponseCode = $request->get('vnp_ResponseCode');//Mã tham chiếu của giao dịch 00==>thanh cong
+        $vnp_TxnRef = $request->get('vnp_TxnRef');//Mã đơn hàng
+        $vnp_Amount = $request->get('vnp_Amount');//Số tiền thanh toán
+        // Kiểm tra data xem kết quả giao dịch có hợp lệ không
+        if($vnp_ReponseCode != null){
+            if($vnp_ReponseCode == 00){
+                Cart::destroy();
+                return redirect('checkout/result')->with('notice','Thanh toán đơn hàng thành công!Vui lòng kiểm tra email');
+            }
+            else{
+                //Xóa đơn hàng trong database
+                Order::where('id',$vnp_TxnRef)->delete();
+                return redirect('checkout/result')->with('notice','Thanh toán đơn hàng không thành công! Có lỗi xảy ra!');
+            }
+        }
     }
     public function show(){
-        $notify = session()->get('success');
+        $notify = session()->get('notice');
         return view('Frontend.checkout.show',compact('notify'));
     }
 }
